[32] - refactor namings, fix tests, include lot main image url in response
