+++ src/Auctionify.API/ClientApp/angular.json
@@ -101,5 +101,8 @@
         }
       }
     }
+  },
+  "cli": {
+    "analytics": false
   }
 }
