+++ src/Auctionify.Application/Features/Lots/Queries/GetAllLotsWithStatusForSeller/GetAllLotsWithStatusForSellerResponse.cs
@@ -1,6 +1,6 @@
 ﻿using Auctionify.Application.Features.Lots.Queries.BaseQueryModels;
 
-namespace Auctionify.Application.Features.Lots.Queries.GetAllActiveForSeller
+namespace Auctionify.Application.Features.Lots.Queries.GetAllLotsWithStatusForSeller
 {
 	public class GetAllLotsWithStatusForSellerResponse : GetAllLots
 	{
